init sphinx docs
This is setup with the initial sphinx docs, generated using `sphinx-quickstart`
To build docs, run the following
`pip install sphinx`
`cd docs`
`make html`

and that will run the build process for the docs, which should create html pages within `docs/build/html`. This can be set down the line to be run in a github action and pushed to a separate branch for official deployement.