feat: add github workflow for docs deployment
This workflow is a standard github action set to run on any commits/pull requests to `main`
Additionally, the required packages for development needed to be included in the `setup.py`. This should eventually be moved to a more modern `pyproject.toml` file which allows for specifically differentiating the pip install between basic, dev, docs, etc. but for now this should be fine.